refactor: flask conversion for collection endpoints
