+++ src/pyflask/collections/__init__.py
@@ -1,0 +1,7 @@
+from .collections import (
+    get_all_collections,
+    get_current_collection_tags,
+    upload_collection_tags,
+    remove_collection_tags,
+    upload_new_tags
+)
